Se agrego el update
Se agrego el modal el controlador y se checo la base de datos.

+++ database/migrations/2022_12_11_001059_infokids.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('course');
             $table->string('address');
-            $table->integer('movile_no');
+            $table->bigInteger('movile_no');
             $table->timestamps();
         });
